requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-openai
langchain-google-genai
pydantic
python-dotenv
twilio
httpx


